12/5/97 Mike Kaufman
Following additions to Profiler.h made today:
--------------------------------------------

1)  Added public inline void FunctionInfo::IncrNumSubrs(void).  This
funciton is similiar to the IncrNumCalls() method of FunctionInfo.

2)  Added InclTime_cs and ExclTime_cs to *both* the Profiler and the
FunctionInfo objects.  These fields are used in the
Profiler.CallStackTrace() function to calculate callstack-respecitve
inclusive and exclusive times for both profilers and their functions.  I
flagged these edits with "mrk" or "mike" or something if you need to find
them quickly.

3)  I removed the '++NumSubrs;' line from FunctionInfo::ExcludeTime().
This statement was encompassed in the function described in 1) above and
is not called from Profiler::Start() method.

What needs to be done:
---------------------

1)  I need to #define PROFILE_CALLSTACK and place that around various
sections of code to comply with the current methodology of only compiling
necessary features of the Profile.o.   This also needs to be done in
Profiler.cpp.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@19 a8e31b63-e6ca-409b-a01b-2eec1b506579