Added Pooma Tracing code to TAU standalone profiling.
TAU Tracing package integrated in POOMA for generating traces for VAMPIR.
*************************************************************************
Changes:
1) Tracing can exist with or without profiling. FunctionInfo was extended
with member FunctionId (index to the FunctionDB vector) and has corresponding
GetFunctionId method.
2) Extended FunctionInfo ctor to have a text string of the primary profile group
(e.g. TAU_FIELD | TAU_UTILITY would get "TAU_FIELD" as a string stored in
string GroupName. GroupName is stored only when tracing is enabled.
3) Added macros TAU_TRACE_SENDMSG and TAU_TRACE_RECVMSG to trace messages
(defined to null if -DTRACING_ON is not defined). Corresponding routines for
RtsLayer::TraceRecvMsg and RtsLayer::TraceSendMsg added.
4) Profiler::Start now triggers a function entry event to be traced as does
Profiler::Stop a function exit (This allows both timers and functions to be
traced).
5) Profiler::StoreData() triggers the close of trace file and RtsLayer::DumpEDF
routine (added for tracing) which stores the event id table (<eventid, event name> tuples). An event name could be a function/timer name as well as an external
event such as INIT event, message send/recv event or FLUSH on trace.
6) RtsLayer::setMyNode() triggers the creation of a trace file as it is the
earliest time when we know the rank of the process. The events that take place
prior to TAU_PROFILE_SET_NODE invocation are buffered and their nodeid is
corrected to reflect the correct value before being traced.
7) Environment variable TRACEDIR is used to find the directory to store
event table (.edf) files (one per node) and trace (tau.<node>.trc) files.
A default value of "." is used if it is not defined.
8) To selectively trace certain functions, use commandline option --profile
as is the case with profiling. All other profiling options can be used for
tracing too e.g., -DSGI_TIMERS on O2K's to access fast nanosecond timers
instead of paying for gettimeofday() for getting time.
9) The traces are currently generated in the pC++/TAU trace format (originally
written by Pete Beckman and Bernd Mohr). The issues of dynamic name registration
for templated functions have been incorporated in this release in the form of
generating unique event definition files for each node for those functions and
events that were invoked.
10) Tested with Vampir on SGI64.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@28 a8e31b63-e6ca-409b-a01b-2eec1b506579