Made field iterations global. [Ref: Bernd, Hongyon (PSU)].
Date: Thu, 2 Aug 2001 09:11:29 +0200
From: Bernd Mohr <b.mohr@fz-juelich.de>
To: Hongyon Suauthai <suauthah@cs.pdx.edu>
Cc: tau-bugs@cs.uoregon.edu
Subject: Re: TAU in openmp

hi,

> I have successfully linking TAU with RWCP Omni compiler but I have
> questions regarding the example OpenMP program in c.
>
> 1. There is a typedef for an int[][] in ppm.h call field.
>    Everytime I ran the program, there was a segmentation fault before
> it could even start the parallel section. So I put the variable "field
> iterations" as a global variable and make it as a threadprivate.
> This runs correctly.  I am not sure why?? (Could it be a bug in Omni
> compiler itself?)

I assume you talk about the declaration of "field iterations;"
in function main() in file "mandel.c".

If "iterations" is declared as local variable of main(), it is stored
on the stack. Depending how the OpenMP compiler implements the OpenMP
functionality. it can happen that main() is executed in each thread.
If this is the case, "iterations" is then actually stored in the
stack for each thread which is typically much smaller than "general"
(sequential) stack. The effect is that you get a stack overflow which on
Unix typically results in a segmentation fault.

Solutions:
1) make iterations global as you found out by yourself; this way it is
stored in static memory and not on the stack.
2) increase the thread stack size; this depends on the compiler / thread
system and I don't know how to do this for Omni.


Bernd

--
Bernd Mohr / Research Centre Juelich, ZAM, Germany / B.Mohr@fz-juelich.de


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@846 a8e31b63-e6ca-409b-a01b-2eec1b506579