From: Julian C. Cummings <cummings@cacr.caltech.edu> To: Sameer Suresh Shende <sameer@cs.uoregon.edu> Subject: RE: new version of TAU released
Sameer,

There is a cut-and-paste error in the configure script.  In the
part that handles the -fortran option and the case of absoft,
it is adding IBM64_FORTRAN to the list of fixmakefile args.
It should be ABSOFT_FORTRAN instead.  After I fixed this,
everything looked good with the new -fortran and -mpi options.

Julian C.



I spoke a bit too soon on the -fortran=absoft option,
because there is another problem.  Since Absoft is not
a native compiler, you really need a -L flag to specify
where to find the Fortran libraries.  Any ideas on how
to do this?  This is not a problem for our build procedure,
since we assume that ABSOFT_LIBDIR is defined and tells
us where to find the libraries.  But I tried to build the
Tau f90 example and it failed because the libraries are
not found.

Julian C.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@942 a8e31b63-e6ca-409b-a01b-2eec1b506579