Added TAU_LDFLAGS to the libraries line [Julian].
The latest version of the configure script works well on IBM.
There is one further refinement I would suggest with respect
to the special MPI flags on AIX.  You should put the
-binitfini:poe_remote_main flag into a separate macro such
as TAU_LDFLAGS rather than adding it to TAU_MPI_LIBS.  The
reason is that as I noted previously, this flag is to be used
when linking parallel executables, but not shared libraries.
In this regard, it is distinct from the MPI libpath and library
flags.  If you put it into a separate macro, the user can then
discriminate between these two situations.  For example, in the
Makefile for the "pi" example, you would want to add TAU_LDFLAGS
to the definition of LDFLAGS in order to include this flag.  I
think the current setup won't quite work when I go to build our
code because -binitfini:poe_remote_main should not be included
when building shared objects.  (Our code consists of several
shared objects that are dynamically loaded into Python.)  I can
work around this problem for now.

Julian C.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@953 a8e31b63-e6ca-409b-a01b-2eec1b506579