From: Julian Cummings <cummings@cacr.caltech.edu> To: Sameer Suresh Shende <sameer@cs.uoregon.edu> Cc: Robert Micha Ansell-Bell <bertie@darkwing.uoregon.edu> Subject: Re: a few SGI glitches
Hi Sameer,

I made a fairly trivial change to the RtsLayer.cpp code
which got me what I wanted, namely a reasonable default
behavior if PAPI_EVENT is not set.  In RtsLayer::USecD(),
just below the line #ifdef TAU_PAPI, I added

  static const char *papi_env = getenv("PAPI_EVENT");
  if (papi_env != NULL)
    return PapiLayer::getCounters(tid);

This just checks if PAPI_EVENT is set and calls getCounters()
if it is.  If not, we still can call getWallClockTime() or
getVirtualTime() if Tau was configured with -PAPIWALLCLOCK
or -PAPIVIRTUAL.  If neither of these configuration options
was used, we will call getCounters() and issue the error message
saying that PAPI_EVENT must be defined.

Once you modify your interface to handle multiple counters, you
can probably just make the PAPI wallclock time or virtual time
the zeroth counter entry by default if Tau was configured with
-PAPIWALLCLOCK or -PAPIVIRTUAL.

Regards, Julian C.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@1005 a8e31b63-e6ca-409b-a01b-2eec1b506579