In determining a routine's return type, we use fullName instead of name. This fixes the problem when a C++ file is instrumented as a C file (START/STOP instead of TAU_PROFILE to lower overhead of disabled groups [CACR]).                 {                   const pdbType *t = ((pdbRoutine *)((*it)->item))->signature()->returnType();                   if ( const pdbGroup* gr = t->isGroup() )                   {                     return_string = gr->fullName();                   }                   else                   {                     return_string = t->fullName();                   }                 }
Hi Sameer,

Those changes in tau_instrumentor.cpp fixed the problems
for all but one of our C++ source files.  I am attaching the
files for the one piece of code that still does not instrument OK
with C-style instrumentation.  The basic problem is that the
original function returns a qualified type classname::iterator,
but in the instrumented code, tau_ret_val is declared as "iterator"
without the classname qualifier.  Look at _splitEdges(), for example.
I think this is the only problem left to fix.

Thanks, Julian C.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@1013 a8e31b63-e6ca-409b-a01b-2eec1b506579