For IBM, we use libpapi.a not .so. Reason: Date: Mon, 25 Feb 2002 18:07:00 -0800 From: Julian C. Cummings <cummings@cacr.caltech.edu> To: Sameer Suresh Shende <sameer@cs.uoregon.edu> Hi Sameer,
I hope you had a nice vacation.  There were still some minor
glitches when using Tau with papi on AIX.  You added the -rpath
flag for linking with a shared version of libpapi, but that
isn't correct for the IBM linker.  So I think you need to
remove that from the IBMPAPI line in the Tau Makefile.

Trying to use the shared version libpapi.so turned into a
nightmare for several reasons.  First, they don't use the
correct loader options to build it.  Once I corrected that,
I discovered that their library references the _etext symbol.
This is the address of the end of the text section of the
executable, and it is defined by the loader.  In order to
avoid undefined references, you have to add special loader
flags to tell libpapi.so that it will import the symbol
and tell the executable to export the symbol.  Kind of
annoying to put this extra burden on an executable that
wants to use libpapi.so, but symbols are not automatically
exported on IBM.  This was sufficiently annoying that I
ended up switching back to using the static archive libpapi.a.

Here's the kicker: papi by default tries to build both the
shared and static versions of the library and puts them in
the same directory.  So if you specify the directory with a
-L flag and then use -lpapi, you may get one version or the
other, depending upon whether you are building a shared object
or an executable.  I had to specify libpapi.a with the full
pathname in order to ensure linking with the static version.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@1051 a8e31b63-e6ca-409b-a01b-2eec1b506579