Date: Tue, 12 Mar 2002 14:37:26 -0800 (PST) From: Leslie Zack Blanke <lblanke@mailhost.cs.uoregon.edu> To: sameer@cs.uoregon.edu Subject: Re: Path
Sameer:

Alright I made those changes.  If no output file is specifed then .prv is sent
to stdout, and the .pcf file is created with a default name "config.pcf".
This really makes no difference because the .prv file was never created thus
can
never be loaded into paraver and so will not require a .pcf file.  If the name
given as the output is not atleast 5 letters long  e.g.(-.prv) then we tell the user that the outfile must be of the form *.prv and then exit.
"Stacked Val" is used as the default compose function".  And the header will
always indicate one processor running 1 PTask with 1 task, with n threads
running on the 1 processor as we discussed.  I don't seem to have or can't find tau2 directory, so I set permissions so that you should be able to access
tau_convert in /research/cluster/scratch1/lblanke/Tau/tau-2.10.11/utils

Let me know if there are any other changes that you see need to be made and I
will take care of them right away.  Sorry this happened in the first place.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@1057 a8e31b63-e6ca-409b-a01b-2eec1b506579