When groups are defined, original code was:     if ((it = funcDB.find((const char *)functionName)) != funcDB.end()) { #ifdef DEBUG       cout << "Found the name " << functionName << endl; #endif /* DEBUG */       delete functionName; // don't need this if its already there.     }     else       funcDB[(const char *)functionName] = FunctionData();       // adds  a null record and creates the name key in the map       // Note: don't delete functionName - STL needs it         /* PROCESS NO. OF Invocations Profiled */
    //Just added functionName, therefore it will be there.
    //Set the group names for this function.
    if(groupNamesUsed){
      char *createdGNSpace = new char[strlen(groupNames)+1];
      strcpy(createdGNSpace, groupNames);
      funcDB[functionName].groupNames = createdGNSpace;
    }

This is incorrect. When the name does not exist in the FuncDB, note functionName is deleted -- it can't be used in funcDB[functionName].groupNames = createdGNSpace! The entire if clause has to be moved into the else branch. So, { } were
added to correct this problem. This error was noticed in multi-threaded profiles
where we got an error message:
Reading Profile files in profile.*
ERROR : In second pass ProcessFileDynamic didn't find name first() int () in file profile.0.0.0


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@1177 a8e31b63-e6ca-409b-a01b-2eec1b506579