Modified tau_run.cpp so it will now work on solaris and sgi.  In the invokeRoutineInFunction() in tau_run, I added a simple check that makes sure that BPatch_Vector of points that is passed to insertSnippet() is not null. Previously, we had been getting a segmentation fault on both platforms.
After this problem was taken care of, I proceeded to install the dyninst API on
on the SGI and Solaris platforms. On the Solaris OS, I needed to compile it
with the same version of gcc that they used, gcc-2.95.3  Using this version of gcc, I was able to successfully compile and run the tests included with
the dyninst library package.  However, when I tried using tau_run, there were
some cases where the mutatee would end early, and therefore a profile file
would not be produced.  When I ran the klargest program with no arguments,
tau_run would run successfully and produce a profile file.  However, when I
used the argements 1000000 and 2324, the program would end prematurely.  I
tried using different compilers, but the problem persisted.  At this point, I
created several dummy programs to test tau_run on.  The first was a simple
hello-world program, and it finished successfully for both compilers.  Next, I
tried a simple program that called a trivial function.  This program also
produced the profile file via tau_run.  Next, I modified the previous program
so that it would call the trivial routine 10,000,000 times.  This also worked
correctly.  Next, I used the UNIX top tool to watch the memory usage as I
ran tau_run with klargest 1000000 2324.  It appears that the problem was
that the mutatee was stopping because it ran out of memory.

Next, I wanted to run the same tests on the SGI machines, because they had a
lot more memory available. This time I compiled the dyninst API with
gcc-2.95.2, and it compiled correctly except for part of one of the tests.  I
also had to add the libiberty.a library to the
dyninstAPI-3.0/lib/mips-sgi-irix6.5/ directory.  All the compiled tests ran
correctly.  Now, with the extra memory, tau_run klargest 1000000 2324 was able
to run to completion and produce the expected profile file.  Also of note,
the test program that called a trivial routine was producing an unexpected
profile file, but this turned out to be because of compiler optimization, and
the unneccessary routine was dropped.

-Nick


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@1256 a8e31b63-e6ca-409b-a01b-2eec1b506579