Fixed the bug that tracked the multiple counter name. The problem was as follows:
static char * counterName = NULL;

bool IsDynamicProfiling (char *filename)
{
  char version[64];

  // read version from filename
  counterName = version;

// version goes out of scope -- destroyed/becomes NULL/undefined
}

int main() {
// ...
            printf ("default.dep\n%d %s\n", numfunc, counterName);

}

As a result, counterName contains garbage because it *points* to (and not
contains the contents of version when it was valid). The fix:
counterName  = version;
is changed to
counterName = strdup(version);

so it does a malloc and *copies* the contents of version in it. Now it works properly.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@1295 a8e31b63-e6ca-409b-a01b-2eec1b506579