Julian's fix to ignore trailing whitespaces in the exclude list names: Hi Sameer,
OK, I figured out why the modified tau_selective code was failing to
exclude some
functions that were in the exclude list.  The problem was that the new
wildcardCompare
function did not properly ignore trailing whitespace in the exclude list
names.  I fixed this
problem by changing line 293 of tau_selective.cpp to read

   while ((*wild == kleenestar) || (*wild == ' ')) {

This will skip over any trailing whitespace in the name to be matched so
that a proper
match is still obtained.  The old code handled this with the
size_to_compare argument.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@2299 a8e31b63-e6ca-409b-a01b-2eec1b506579