Earlier, we had to be very careful in how we invoked tau_instrumentor. If we'd generated a PDB file using: % cxxparse /home/pkg/foo.cpp and it contained so#1 /home/pkg/foo.cpp then to instrument the file, we *had* to invoke tau_instrumentor as: tau_instrumentor foo.pdb /home/pkg/foo.cpp -o foo.inst.cpp the same applied to the matching of file names ./foo.cpp and foo.cpp.
Now, with a fuzzyMatch function, we can do better! It allows us to match such
names that string operator (a == b) would normally classify as a mismatch.

Also, if none of the files in the PDB match our current file, we still copy
it out as output file name so that the make can proceed.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@2387 a8e31b63-e6ca-409b-a01b-2eec1b506579