Dyninst v4.1 has a bug: we can't instrument main's entry under Linux. This fixes the bug by calling TAU's top level routine stub under Linux and Dyninst 4.1. Arrrgh! I hate such stop gap bandaids. Date: Thu, 17 Jun 2004 16:12:38 -0500 (CDT) From: Matthew Legendre <legendre@cs.wisc.edu> To: sameer@cs.uoregon.edu Cc: bugs@dyninst.org Subject: Re: DyninstAPI bug report
On Wed, 16 Jun 2004, Sameer Shende wrote:
> Hi,
>       I'd like to report an error in instrumenting main under Linux. I
> can instrument its entry point under Solaris but not under Linux. I've
> attached the relevant makefiles and code. I'm using RedHat Linux 8.0
> (Kernel 2.4.20-28.8smp) and SunOS 5.9.
>       I'd appreciate any pointers/suggestions to work around this bug.
>       Thanks!
>       - Sameer


I believe the problem here is that when the mutatee is loaded, Dyninst
will immediately run it until it reaches the first instruction of main
(for initialization purposes).  By the time you insert instrumentation
into the entry point of main, the program is already passed the
instrumentation point.  In fact, you're probably going to have trouble
running instrumentation within the first 5 bytes of main.

This is a certainly a Dyninst bug.  I'll add it to our database and
hopefully someone will take care of it soon.  In the mean time, would it
be possible for your program to initialize itself in a way that assumes
entry instrumentation for main has already been run?

-Matt


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@2404 a8e31b63-e6ca-409b-a01b-2eec1b506579