Added GetFlushEvents(int tid) which returns 1 if events.<n>.edf file needs to be flushed (this happens when a new event is recorded).
Also, in Java, thread 1's termination has the new node id (non-negative) and
this event triggers the trace file creation. This happens in Java 1.4.1.
So, when EvInit is called the first time, it doesn't create the trace file (as the nodeid is -1), but when it is called the second time, it does create a
trace file and adds the WALLCLOCK time record and returns the no. of records created (1). In our algorithm to treat these new records, we were initializing the
timestamps of all events from 0 to n-1 (when n records are created). This is wrong. We need to initialize the timestamp of only the n records before the current event. So, instead of going from 0 to n-1 setting timestamps, we go from
currentevent - 1 - 0 to currentevent - 1 - (n-1) setting the timestamps equal to the current timestamp.  This is done so time does not go back in the trace file.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@2676 a8e31b63-e6ca-409b-a01b-2eec1b506579