Paraprof crashes if you load it in a directory that contains a directory named "profiles". I would get:
java.io.FileNotFoundException: /mnt/disk2/taudata/miranda/16k/profiles (Is a directory)

I've fixed part of this problem, but a larger part still remains.  It
seems that once it has decided that there are profiles in a directory
(even if there aren't), there's no good way to back out of it.  The
DataSession.initialize method doesn't even have a return code.  We'll
have to revisit this later.  Currently, the rest of ParaProf expects
profiles to be there, so it crashes in some drawing routines.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@2769 a8e31b63-e6ca-409b-a01b-2eec1b506579