Added -exec-prefix=<dir> option. [Ref: Shirley, OSPAT]
Date: Mon, 6 Dec 2004 16:35:42 -0500 (EST)
From: Shirley Moore <shirley@cs.utk.edu>
To: Dan Terpstra <terpstra@cs.utk.edu>
Cc: OSPAT-general <ospat-general@cs.utk.edu>
Subject: [OSPAT-general] [INSTR] [PROF]

I would also like to suggest INSTALL for a common directory hierarchy
and convention for where to put files (e.g., header files, libraries, an
executables) when installing tools.  This may or may not be worth a SIG
but it would be nice to agree on it somehow.  What I have in mind is
being able to specify a directory (e.g., -prefix=<prefix-dir>) for
platform-independent files (e.g., header files) and a directory
(e.g., -exec-prefix=<exec-prefix-dir>) for platform-dependent files (e.g.,
libraries and executables) when installing a tool.  Then the tool would
look for or create an include directory in <prefix-dir> and look for or
create bin and lib subdirectories in <exec-prefix-dir>.  And other tools
could assume that files for other tools they depend on are installed in
these locations.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@2874 a8e31b63-e6ca-409b-a01b-2eec1b506579