Fixed stat window to immediately update when group selections are made.  Previously, you would have to open a new stat window to see only the groups you want.
Changed %Time -> %Total Time (and %Counts -> %Total Counts) as recommended below:

Also changed "Select Value Type" on the stat window to be "Sort By",
to prevent confusion as below:


Date: Wed, 01 Dec 2004 16:41:28 -0600
From: Avi Purkayastha <avijit@tacc.utexas.edu>
To: Alan Morris <amorris@cs.uoregon.edu>
Cc: Sameer Shende <sameer@cs.uoregon.edu>
Subject: Re: Difference in time between pprof and paraprof

Alan,

>
> The %Time shown in the "Show Mean Statistics Window" is always the
> inclusive percentage (as in pprof).  When you left click on a thread or
> the mean, the exclusive percentage value is first shown.  This value is
> not available in pprof.  Go to "options->Select Value Type->Inclusive" to
> see inclusive values.

Makes sense, now that you explain it. However, what is confusing to me is,
as you said "Show Mean Statistics Window" is always the inclusive percentage
(as in pprof), but the "Sorted by" category is showing exclusive or whatever
we choose, the Time% remains inclusive time, which is what seems to me to be
contradictory. If you put in parenthesis of %Time (inclusive) then it would
be clear that "options->Select Value Type->whatever", then the Time% is
still inclusive no matter which option is chosen as a sorted value.

My 2c to update/fix or document.

    -- Avi


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@2907 a8e31b63-e6ca-409b-a01b-2eec1b506579