1) Fixed MpiP reader: inclusive per call value was in milliseconds instead of microseconds.  I previously fixed this for inclusive/exclusive times, but didn't notice inclusive per call.
2) Massive reworking of error handling throughout the API.  Gone are the dozens
of try/catch blocks at the start and end of almost every function.  Now the
exceptions are actually used for what they were meant for, actually handling
the error (rather than just printing a message to stderr).  This is an ongoing
process and is not complete, but the groundwork is there.

3) When any part of saving a trial to the database fails, this will trigger a
rollback so that all of the previous inserts will be aborted.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@2915 a8e31b63-e6ca-409b-a01b-2eec1b506579