Fixed Gprof reader: 1) Units, milliseconds instead of microseconds 2) Number of subroutines (children), now we read the numerator in each child line instead of just adding up the number of children. 3) Inclusive per call is now calculated and stored.
Fixed Pprof reader, total numSubr, numCalls, and inclusivePerCall were not being read

Fixed HPMToolkit for changes in Function/FunctionProfile

Removed NCT and TrialData objects.  There was always a one-to-one mapping with
them and the DataSource, this also removes a whole bunch of useless
DataSource.getTrialData() and DataSource.getNCT() calls.

Trimmed Function and FunctionProfile objects down in size.  Heap usage for
Miranda16k is now 300,942kb vs 324,132kb in TAU 2.14.  This is despite the fact
that we now store the number of subroutines and calls as doubles, a change that
increased usage to 337,558kb.

Started cleaning up constructors for DataSource objects.  Instead of an Object,
they will now take something more in line with what they use.  This has been
done for GProf (now an array of File's).


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@2947 a8e31b63-e6ca-409b-a01b-2eec1b506579