Rewrote uploading of full trials to allow cancelling (with rollback support). Now it now doesn't bother creating IntervalEvents/IntervalLocationProfiles/etc, which means it will require roughly 1/2 the memory to upload large trials.  It is now trivial to change to batch uploads, but it is being left disabled until it can be fully tested with all the databases.  It's not clear that all JDBC drivers even implement it anyway.
There still seem to be some unresolved issues with partial trial uploads.  Even
with autocommit turned off, if you exit the program, the data uploaded so far
gets committed.  We should be able to fix this for the most part but I have
read numerous accounts where transactions get committed even when the JVM
crashes, which we can't possibly do anything about.  Further analysis is required.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@3015 a8e31b63-e6ca-409b-a01b-2eec1b506579