Lots of changes:
1) Changed the way "selected metrics" work so that you can view data from
different metrics for the same trial at the same time.  Window's can now choose
Exclusive->PAPI_FP_INS, etc.  Currently only the FunctionDataWindow and the
CallGraphWindow fully implement this functionality.

2) Changed the way preferences work and combined them with the colormap.
Preferences are now global instead of per trial.  Now the user can change the
color of MPI_Recv on one trial and see the change in all trials.  A new window,
the Color Map Window, was added which displays a list of "assigned" colors that
will be applied to all trials.  This is separate from the "default" colors
which are now stored with the preferences file.  These are not one in the same
since the primary use of the colors is to show distinction among functions.
Thus, except for "assigned" colors, the defaults will be loaded in mean value
order such that we do our best to never have two of the same color next to each
other.  This window also allows you to "assign" all of the colors from
functions from a given trial to the assigned list, thus syncing all other
trials with respect to common functions (this is useful for phase based
analysis).

3) Added exclusive per call values (only on the more advanced metrics
displays).  It will soon be applied to the other windows.

4) Added load/saving of preferences files.  Additionally, the preferences are
now saved on exit by default.  So color/font changes that have been applied
during a session will persist.

5) The paraprof manager window now "remembers" it's position.  This is a boon
for dual-monitor users who will no longer have to deal with the window always
popping up between the two screens.  Java 1.5 supports functionality to allow
the native window manager to place the window, but alas, 1.5 is too advanced for
our purposes.

6) Other changes that I can't remember right now.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@3086 a8e31b63-e6ca-409b-a01b-2eec1b506579