Lots of changes and additions.
1) Finally fixed the Thread Data Window such that the horizontal scroll bar
corresponds exactly to the longest function name, rather than the "temporary
fix" of +500.

2) Fixed the unintended linking of value type in one function window to
another.

3) Finally got rid of the old "display sliders" method of selecting the bar
widths.  No more drop-down boxes with 'multiples'.  Now just a simple, 100%
smooth slider labeled with actual pixel values.

4) Fixed the units in the 3d displays (calls/#subr were being divided when
using units other than microseconds).

5) Selected function/thread are now sync'd between the bar plot and triangle
mesh plot.

6) Added transparency (actually, translucency) modes to the bar plot and
triangle mesh plot.  The selection style changes as well (slices to triangle
mesh, full highlighted bars to the bar plot).

7) Bar plot scalability has been vastly improved.  This was done with smart use
of display lists and a bit of cheating (currently a threshold exists such that
values too small are not given a box, this has the added bonus that it improves
visibility of small values not below the threshold).  There is something to be
investigated about the difference between "no value" and 0.  The bar plot now
scales to 64k processors (derived from miranda data), I ran out of memory
trying to do 128k.  We'll have to try it on optix.

8) Bar plot allows for 8 way translucency (so that boxes are rendered in the
proper order for every view direction), but as such it requires 4 times the gfx
memory and is not recommended for massive datasets.  The triangle mesh plot
only allows for 1-way translucency, so depending on which way you look through
it, it may not be as impressive (the selection slice & axis will always show
through, this only applies to other translucent objects).

9) Added a "reverse video" option to the render control panel, giving a black
on white view that might be more conducive to projection displays (and printers).

10) Added several (hidden) command line options, mostly for fun (-Dvis.fps to
show fps during rotation, -Dvis.polyfill to force glPolygonMode(...) to
GL_FILL, -Dvis.polyline to GL_LINE, and -Dvis.polypoint to GL_POINT.  Also
added is -Dparaprof.verbose to provide some informative messages.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@3151 a8e31b63-e6ca-409b-a01b-2eec1b506579