Added preference option to reverse callpaths (C <= B <= A rather than A => B => C).
Increased maximum size of bar/trianglmesh plots to better accommodate larger
numbers of functions.

Reworked some pop-up menus to make them more consistent (and easier to
maintain).

Added ParaProf Packed Profile format (PackedProfileDataSource.java), a
portable, binary, highly efficient and gzip compressed data format for storing
profiles.  Tentatively with the extension .ppk.  The miranda 16k set is 11MB as
a ppk file and loads in 10 seconds on my machine (down from over a minute as
tau profiles).  This format was added to address numerous issues:

First, there is no export functionality.  There was once an XML schema, but it
was extremely poorly thought out.  Besides, goal number two was *small size*.

A new format had to be created because none of the supported formats had all
the features of the others, this is a superset/union off all the other formats.

Accordingly, a new option is available in the right click menu of a trial to
export to this format.  This is currently the only way to get data out of the
database (and thus move it to another database).

This format (obviously) needs a lot of testing before it is ready for release.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@3195 a8e31b63-e6ca-409b-a01b-2eec1b506579