3D window now re-sorts when "reverse callpaths" is checked.  In the future I will add options to choose sort methods other than "by name" for this window.
Packed format now has header information to protect against newer versions.  A
compatibility id exists to allow for some degree of backward compatibility (so
that if we add trial metadata to the format in the future, the current version
will still be able to read the files, it will just skip over the metadata).


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@3205 a8e31b63-e6ca-409b-a01b-2eec1b506579