Fixed alignment issues for 3d windows when names are very long (large callpath profiles).  It turned out to be a bug in GLUT that was only recently fixed in GLUT 3.8.  The code was summing up character widths using an int rather than a float, so it ended up underestimating the width of strings.
Full text of the bug:

The deprecated glutStrokeWidth and glutStrokeLength routines return an integer
rather than a float. Newer GLUT programs should use the newer float-returning
routine glutStrokeWidthf and glutStrokeLengthf. Note that these newer float
routines do not exist in implementations prior to GLUT 3.8. Also, the
glutStrokeLength routine summed integerized character widths so it
underestimated the length of strings prior to GLUT 3.8, but the implementation
of glutStrokeLength in GLUT 3.8 and beyond sums the character widths for the
strings in a float (though glutStrokeLength still returns an integer).

---

I have patch GLUT.java in the JOGL source tree and submitted a patch to the
JOGL developers, but I do not have the resources to rebuild the combined
multi-platform jogl.jar, so instead I have imported the necessary files into
our project until the next release of JOGL.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@3308 a8e31b63-e6ca-409b-a01b-2eec1b506579