Added standard deviation as a regular item (like mean).
Ctrl-Click now does the same thing as right-click (for those unfortunate mac
users with only one button).

Added toplevel menu items to access most of the windows.

Added ability to change meaning of "mean" (and std. dev.).  Mean can either be
the total divided by the number of threads, or the total divided by the number
of threads which call the function in question.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@3337 a8e31b63-e6ca-409b-a01b-2eec1b506579