Refactoring time...
Modified the R code to work in the same framework as the Weka code, and
made the code engine-specific.  The AnalysisTaskWrapper.java file has
been replaced with AnalysisTaskWrapper.cpp, and the Makefile is modified
to use CPP to pre-process the .cpp file and output the .java file.  The
only difference is whether R or Weka is used as the engine.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@3647 a8e31b63-e6ca-409b-a01b-2eec1b506579