For MULTIPLECOUNTERS GetInclTime does a malloc of a big array and returns the full array. This is fine for GET_FUNC_VALS kind of routines, but when we want to measure the effect of throttling, it is an unnecessary overhead - not to mention a huge memory leak! So, we introduce GetInclTimeForCounter  - a function that takes the tid and a counter id and returns the double incl (or excl) time.
What is interesting is that throttling only works when PROFILING is turned on.
So, to use it with tracing, you need to enable both profiling and tracing. This
is because we do not store the inclusive time. Maybe we should store just the
starttime and calculate it during Tracing. Alan?


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@3830 a8e31b63-e6ca-409b-a01b-2eec1b506579