Fixed perfexplorer configuration so it works with -prefix values.
When the user runs configure with -prefix, an etc directory is
created with the necessary files to configure perfexplorer.

A script is created, called perfexplorer_configure, which has to
be executed before perfexplorer can be run.

The perfexplorer_configure script will configure perfexplorer, and
download any files necessary.  The wget used is the java Wget in
tau-common.jar.

The configure.skel is still used to build a configure script for
building perfexplorer, which is necessary due to the need to
specify the R build environment, if R is used.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@3994 a8e31b63-e6ca-409b-a01b-2eec1b506579