Added an option for preprocessing source code prior to parsing. Using -optPreProcess we can now pre-process the source using /usr/bin/cpp -P. If you need to use a different pre-processor, use -optCPP=<file> and -optCPPOpts=<options> and -optCPPReset=<opts> for setting a preprocessor, adding options or resetting the default options used.
Another big change is the way we used to remove -I and -D elements from the Fortran
command line. The problem was that we could not distinguish the compile options
passed by TAU from those passed by the user or specified on the command line. Now,
we use -optTauCompile=<options> in addition to -optCompile=<options>. This way,
we can specify -optTauCompile in include/Makefile and not pass these options for
Fortran compilation (-fPIC caused havoc to WRF [NOAA]).

Also, -optNoRevert allows the program to end in an error instead of going forth
with the reverting to make error. -optRevert forces revert if one was reset (default).


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@4070 a8e31b63-e6ca-409b-a01b-2eec1b506579