When a single if occurs at the beginning of a routine, the -p perflib instrumentation option did not work properly:
          if(blah) return

      then the instrumented code looks like:

          if(blah)
      then
                call f_perf_update(...)
                return
             endif


	This was because when the rest of the statement is written out at the entry of the
routine, we did not check (for perflib), if print_cr was false before writing out the
carriage return. So, TAU instrumented code did not have this problem, but laperf instrumented
code did. This is now fixed and we check the value of print_cr before writing the carriage
return in the body begin for Fortran code.


git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@4372 a8e31b63-e6ca-409b-a01b-2eec1b506579